Cleaned up the Sound System v1.0.1
Refactored the name of the GameSounds class to a struct named Sounds
(this matches the actual Swift file now).

Put all the sounds in a dictionary that uses the sound name as a key
value.

Added a function that plays the sound and the key feature is that it
returns a message if the sound name is nil.