Made the UI Connection to the Game Stats Page v1.0.6
- Added the code connection to the Game Stats Page. Right now the only
way to get to it is by pressing the restart button, but I’ll fix that
later (I’ll give it it’s own button in the .Ready state).

- The animation isn’t quite right either. I want it to be a staggered
slide in, but the code wasn’t quite working properly. I’ll fix it
(probably just compromise) tomorrow.

- Getting a little late tonight, I’ll finish this up and start making
the calculations for this page tomorrow. Woo hoo! Good day’s work :)