Animation Overhaul v1.0.9
- Completely recoded all the animations in the game.

- The biggest thing in this update is that I completely removed the
SKAnimation files. All animations are hardcoded in the code.

- XCode is great, but the UI to make animations and stuff wasn’t very
flexible and it was a bit buggy at times. Not to mention I’m not a very
good visual worker; this coded animation method is much more readable
and manageable (for me at least)

- Changed the title animation. The title now continues sliding off the
in the same direction it slid on in (prior to this update the top title
label slide on right, then slid off left. Now it slides on right, and
slides off left)

- Changed up the scoreboard a little bit to be more compact. Eventually
I’ll add a settings page where you can set the the scoreboard to the
right or left (this makes it so that left-thumbed players can still see
their score.)

- Tomorrow I’m gonna try and do some retexturing. I suck at art, but
it’s about time I whip out the ol’ Photoshop!