Massive UI Overhaul v1.0.11
- This was a huuuuuuuuge update. Now that my high school’s J-Term has
started, I get 8+ hours a day to work on this project :)

- I fixed up the Game Over Menu a little bit to accommodate a place
where it displays the high score. Later tonight I might make an
animation that slides the scoreboard off the screen when the game ends.

- The Game Stat Page was the biggest change! Instead of making it its
own scene, I made it it’s own little tab that was part of the Infinity
Game Scene.

- If you break a record, the color of the score turns gold.

- The items on the Game Stat Page come in by two’s. Currently, the
share button is nonfunctional, but I’m thinking it will just take a
screenshot and share it to the most common sites (Twitter, FB,
Instagram, even Snapchat if I could figure it out).

- Another huge behind the scenes fix was correcting the Z-Values for
the Infinity Game Scene. I finally set everything so that there isn’t
huge gaps between them. For example, the bunny was 3 and then the
infinity scoreboard was 10. It’s just unnecessary and messy. Either
way, I fixed it all now.