Recoded UI Implementation v1.0.15
- Completely recoded the way I add UI elements on to the screen.

- Previously, I had added all the UI elements onto the screen and just
set this to be hidden. While this wasn’t really affecting performance,
it was extremely messy and with more UI additions it could end up being
much laggier.

- This method of UI implementation also allows for a much simpler
addition of UI elements. Before, it was a really long process with lots
of loose variables and so on. Now it’s easy to manipulate the UI
elements and even easier to add and remove them from the scene (instead
of hiding and unhiding them).

- I’m trying to completely eliminate the InfiniteGameSceneAnimations
class. I’m putting the animations within the UIElements. This is a much
better and more efficient system.

- I also removed the need to use a few timers by using
SKAction.sequences instead of timers and if statements. It’s much more
readable, but there’s no noticeable difference.

- On the GameOverMenu, the labels on the left are always visible and
the values are what come in every 0.5 seconds.