Added GameStats Button v1.0.18
- This is a pretty minor update. I added a button to open up the
GameStats from the startMenu.

- I added a button here that way if a player wanted to show off his
stats he didn’t have to play a game.

- I had to make this GameStats UI different from the UIGameStatsTab
because we didn’t need any of the buttons on this version of the UI.

- I renamed all the SpriteKit Scenes to have the UI prefix. I did this
just to make sure there were no confections between GameStats.swift and
GameStats.sks.

- Later tonight I’ll probably end up adding some better sounds and
possibly music?

- I’ll also be adding the scoreboard switcher to the startMenu
(switches the side of the scoreboard so that left-thumbed players can
still enjoy the game :))