Audio Rework v1.0.19
- Beware, this another extremely messy update… I’m a little rushed
because I’d like to have my demo version done by tomorrow afternoon so
I can work on a short presentation.

- I plan on cleaning everything up as soon as the demo version is done
with much better comments and more Object-Oriented Programming features
(not to mention just more readability and better implementation).

- I added 5 background music songs. The song music-wolf.mp3 is kind of
the games theme. It always plays at start up and during the .Ready
state.

- Once the game starts, a random song is chosen from the five songs and
it fades the intro song out and then fades in the new song.

- Once the hero dies, that song fades out.

- I added a sound panel to the startMenu. If the sound/music is enabled
the button will appear golden. If it’s not enabled it will be white.

- If you disable music, it will fade out the music immediately. If you
enable it, it will fade in immediately (with the intro song).

- All this stuff needs a ton of better implementation. It works, but
it’s all over the place and could easily get bugged, I’m sure. Can’t
wait to clean up this mess :/