Scoreboard Switcher v0.0.20
- First off, this is actually update 1.0.20, but I switched up my
versioning system a bit. Nothing huge.

- I added the scoreboard switcher to the startMenu. This allows you to
switch the side that the scoreboard is on in case you want to use your
left-thumb to click.

- I spaced out all the buttons on the startMenu to be an equal distance
apart.

- This is the official Demo Build for tomorrow’s presentation (J-Term).
This week and the following week will be focused on cleaning up ALL the
code. You’re going to see some major deletions coming up ;)

- I’ll also be adding the share feature in the next few weeks.

- Once everything has been cleaned up and organized, I’ll begin work on
the campaign mode.

- I still need a graphic design ://///