Still Update v0.0.22
I had a weird sync thing happen with GitHub. For some reason, it synced
all my local files with the files from the latest GitHub commit. This
pretty much erased a good day and a half of work, but I was able to
re-add everything back to where it was. I’ll be more careful now and
about accidentally pressing cmd + shift + s in GitHub….