GameStat Recode and Idle Jump Recode v0.0.23
- In this update we totally cleaned up the GameStat class. I thought we
had a good system before, but I realized there would be problems if I
ever wanted to add boolean defaults there would be a problem

- So I ended up just getting rid of my functions to get/stat defaults
because they were pretty redundant. Now I’m just using the default
UserDefaults functions.

- I also changed the Screen Jump methods. Now the bunny does a little
jump into his proper spot on the screen during the .Preparing state.

- I also added back his idle jump thing. I’ll probs end up recoding
this a bit once I clean up the EntityBunny class.

- I had the StartMenu added to the screen during the .Preparing state
that way it reduced the amount of lag that would occur once the
GameState was set to .Ready

- As I’m writing this I’m thinking of a much better system for the
ScreenJump thing. It will be a little difficult to implement, but we’ll
see what we can do :)

- Expect another commit coming tonight.