New Loading Animation v0.0.24
- Instead of having the bunny fall or or jump onto the screen during
the .Preparing state, the bunny slides in on a cloud and just sits
there on his little cloud :)

- In the future, I think I’m going to have him do things like that.
During the campaign, there will be an underwater zone where he will be
incased in a bubble. In the cloud map, he’ll float on a cloud. For the
other levels without a special theme, I think I’ll give him a jetpack
or something lol!

- More recoding tomorrow and throughout next week. Still got a lot of
cleaning to do, but I should be basically done adding new features for
a while.