Scroll Layer Recode v0.0.26
- It’s been a while since I’ve made any progress. Now that semester 2
is in full force I’m finding it hard to find time to do any
programming, but it’s still coming along I guess.

- I recoded the scroll layer system to be A LOT cleaner. Now there’s an
EntityScrollLayer class that has all the scroll functions and stuff.

- It automatically loops the textures infinitely. I’m going to try and
change this up as well and make it a bit more flexible (right now, the
looping point is (0,0), but I have some ideas to make it even easier to
use).

- Next up I’m going to see if I can figure out how to make the obstacle
layer into an EntityScrollLayer. Not sure if I’ll be able to, but we’ll
see :)